refactor: permission 도메인 패키지 구조 리팩터
- domain 패키지의 클래스가 너무 많아져서 하위 패키지로 나눔.
- BaseMember: 불필요한 import 제거
