+++ src/main/java/dev/prvt/yawiki/core/permission/application/PermissionMapper.java
@@ -1,7 +1,7 @@
 package dev.prvt.yawiki.core.permission.application;
 
-import dev.prvt.yawiki.core.permission.domain.Permission;
-import dev.prvt.yawiki.core.permission.domain.PermissionData;
+import dev.prvt.yawiki.core.permission.domain.model.Permission;
+import dev.prvt.yawiki.core.permission.domain.model.PermissionData;
 import org.springframework.stereotype.Component;
 
 @Component
