+++ src/main/java/dev/prvt/yawiki/core/permission/domain/PagePermissionUpdateValidator.java
@@ -1,5 +1,8 @@
 package dev.prvt.yawiki.core.permission.domain;
 
+import dev.prvt.yawiki.core.permission.domain.model.PagePermission;
+import dev.prvt.yawiki.core.permission.domain.model.Permission;
+
 public interface PagePermissionUpdateValidator {
     void validate(PagePermission pagePermission, Permission permission);
 }
