+++ src/main/java/dev/prvt/yawiki/core/permission/domain/impl/PagePermissionUpdateValidatorImpl.java
@@ -1,5 +1,8 @@
-package dev.prvt.yawiki.core.permission.domain;
+package dev.prvt.yawiki.core.permission.domain.impl;
 
+import dev.prvt.yawiki.core.permission.domain.PagePermissionUpdateValidator;
+import dev.prvt.yawiki.core.permission.domain.model.PagePermission;
+import dev.prvt.yawiki.core.permission.domain.model.Permission;
 import org.springframework.stereotype.Component;
 
 @Component
