+++ src/main/java/dev/prvt/yawiki/core/permission/domain/model/AuthorityProfile.java
@@ -1,5 +1,6 @@
-package dev.prvt.yawiki.core.permission.domain;
+package dev.prvt.yawiki.core.permission.domain.model;
 
+import dev.prvt.yawiki.core.permission.domain.AuthorityGrantValidator;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
