+++ src/main/java/dev/prvt/yawiki/core/permission/domain/model/PermissionData.java
@@ -1,4 +1,4 @@
-package dev.prvt.yawiki.core.permission.domain;
+package dev.prvt.yawiki.core.permission.domain.model;
 
 import lombok.Builder;
