+++ src/main/java/dev/prvt/yawiki/core/permission/domain/repository/AuthorityProfileRepository.java
@@ -1,6 +1,6 @@
 package dev.prvt.yawiki.core.permission.domain.repository;
 
-import dev.prvt.yawiki.core.permission.domain.AuthorityProfile;
+import dev.prvt.yawiki.core.permission.domain.model.AuthorityProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
