+++ src/main/java/dev/prvt/yawiki/core/permission/domain/repository/PermissionRepository.java
@@ -1,6 +1,6 @@
 package dev.prvt.yawiki.core.permission.domain.repository;
 
-import dev.prvt.yawiki.core.permission.domain.Permission;
+import dev.prvt.yawiki.core.permission.domain.model.Permission;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PermissionRepository extends JpaRepository<Permission, Integer> {
