+++ src/test/java/dev/prvt/yawiki/core/permission/domain/WikiPageCommandPermissionValidatorImplTest.java
@@ -1,5 +1,6 @@
 package dev.prvt.yawiki.core.permission.domain;
 
+import dev.prvt.yawiki.core.permission.domain.model.ActionType;
 import dev.prvt.yawiki.core.wikipage.domain.model.Namespace;
 import dev.prvt.yawiki.core.wikipage.domain.model.WikiPage;
 import lombok.SneakyThrows;
