+++ src/test/java/dev/prvt/yawiki/core/permission/infra/PermissionLevelConverterTest.java
@@ -1,6 +1,6 @@
 package dev.prvt.yawiki.core.permission.infra;
 
-import dev.prvt.yawiki.core.permission.domain.PermissionLevel;
+import dev.prvt.yawiki.core.permission.domain.model.PermissionLevel;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
